[REFACTOR] redis 관련 기능 한 곳으로 모음

+++ src/main/java/com/out4ider/selleing_backend/domain/gameroom/dto/GameRoomInquiryResponseDto.java
@@ -11,6 +11,6 @@ public class GameRoomInquiryResponseDto {
     private String title;
     private String startSentence;
     private Long roomId;
-    private int maxHeadCount;
-    private int currentHeadCount;
+    private byte maxHeadCount;
+    private byte currentHeadCount;
 }
