+++ src/main/java/com/out4ider/selleing_backend/domain/gameroom/dto/GameRoomRequestDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class GameRoomRequestDto {
     String title;
-    int maxHeadCount;
+    byte maxHeadCount;
     String password;
     String startSentence;
 }
